optimize: revise cloud area related apis.
