+++ src/common/metadata/hostserver.go
@@ -354,5 +354,6 @@ type DeleteHostFromBizParameter struct {
 
 // CloudAreaParameter search cloud area parameter
 type CloudAreaParameter struct {
-	Page BasePage `json:"page" bson:"page" field:"page"`
+	Condition mapstr.MapStr `json:"condition" bson:"condition" field:"condition"`
+	Page      BasePage      `json:"page" bson:"page" field:"page"`
 }
